+++ bug_everyday_ide/mainwindow.ui
@@ -47,7 +47,6 @@
          </property>
         </column>
        </widget>
-       <widget class="QTextEdit" name="EditWidget"/>
       </widget>
       <widget class="QTextEdit" name="ResultWidget"/>
      </widget>
@@ -60,7 +59,7 @@
      <x>0</x>
      <y>0</y>
      <width>887</width>
-     <height>17</height>
+     <height>23</height>
     </rect>
    </property>
    <widget class="QMenu" name="menu_file">
